Replay fromSequenceNr for Akka Persistence event log

+++ src/test/scala/com/github/krasserm/ases/StreamSpec.scala
@@ -39,6 +39,6 @@ trait StreamSpec extends BeforeAndAfterAll { this: TestKit with Suite =>
   def probes[I, O, M](flow: Flow[I, O, M]): (TestPublisher.Probe[I], TestSubscriber.Probe[O]) =
     TestSource.probe[I].viaMat(flow)(Keep.left).toMat(TestSink.probe[O])(Keep.both).run()
 
-  def durables[A](emitted: Seq[Emitted[A]], offset: Int = 0): Seq[Durable[A]] =
+  def durables[A](emitted: Seq[Emitted[A]], offset: Long = 0L): Seq[Durable[A]] =
     emitted.zipWithIndex.map { case (e, i) => e.durable(i + offset) }
 }
